MOD-5647 fix crash on ssl initialization failure (#51)
The PR fixes a crash on hiredis where the SSL initialization fails. It seems that hiredis do not unset the callbacks that was set by the redisInitiateSSL in case of a failure, and so later those callbacks might be called when the SSL context is not initialized and cause a crash.

And issue was open upstream to fix it on hiredis itself: redis/hiredis#1233

In addition the PR add fixes for flaky tests.